Kommentarer og "rengøring" af kode

+++ include/plant_base.h
@@ -5,10 +5,9 @@ class PlantBase
 {
 public:
     virtual float getHeight() = 0;
-
-    virtual void grow(int hours, bool newSoil, double soilMoisture) = 0;
-
+    virtual void grow(int hours, bool newSoil, float soilMoisture) = 0;
     virtual int getMaxHeight() = 0;
+    
 private:
     
 };
